Add CheckboxGroup and RadioCards components
Two new UI components are added in this commit: a CheckboxGroup component and a RadioCards component. These components provide structured groups of checkbox and radio button input UI elements respectively. The code for the CheckboxGroup component was added in a new checkbox-group.tsx file while the code for the RadioCards component was added in a new radio-cards.tsx file. Some edits were also made to the Checkbox, Code, and DropdownMenu components.