Update ESLint settings and remove 'eslint-config-turbo'
This commit involves the removal of 'eslint-config-turbo' from eslint-config's library.js and next.js files, and the renaming of 'react-internal.js' to 'react.js'. The updates to ESLint configurations and renaming of the file are also reflected in the .eslintrc.js of the UI package and the package.json files in both eslint-config and UI package. Changes have been logged in the relevant CHANGELOG.md files, indicating a patch change for both eslint-config (0.0.16) and UI package (0.0.23).