Reorder optional properties and enforce rule for typescript sort keys
Properties within typescript interfaces across multiple files are reordered. The 'typescript-sort-keys/interface' rule in the ESLint configuration has been updated to introduce a mandatory ascending order for keys with 'requiredFirst' set to true. This helps in enhancing the readability and maintainability of typescript code. Version updates on 'eslint-config' and 'ui' packages are also included as per these changes.