Update dialog animation and multiple dependency versions
The code changes include an adjustment to the dialog animation and updates to multiple dependencies in the pnpm-lock.yaml. The animation effect of the dialog box when it opens and closes has been modified, while the versions of several dependencies have been upgraded due to bug fixes or new feature implementations.