Reorganize and refactor UI components in stories
Refactored UI components such as dialog, breadcrumb, alert dialog, sheet, etc in the storybook demo and rearranged their directory structure for better organization. Changes include reordering of class assignments, the addition of new properties, and refinement of existing ones to enhance appearance and functionality. Also bumped the version of the UI package to 0.0.65.