Refactor codebase to rename 'ref' to 'forwardedRef'
Several components in the codebase have had their 'ref' prop renamed to 'forwardedRef'. This change was implemented to make the code easier to understand and maintain as it differentiates between a component's inner ref and a ref forwarded from a parent component. In addition to renaming 'ref' to 'forwardedRef', the commit also made some modifications in the respective components where these props are used.