Rename `slide` to `slideIndex` in slideshow logic
The naming convention in the slideshow logic has been updated to better reflect the content. `slide` has been changed to `slideIndex` in order to clarify that it refers to the index of the slides, not the slides themselves. This change has been carried out across multiple methods in the Vegas Slideshow feature.